Revert "refactor: partial refactoring and compliance with new rules  (#349)" (#350)
This reverts commit 57280ffebcab52256588188f28f71c56a829883d.

+++ system/library/games/crmp/action.php
@@ -16,11 +16,6 @@
  * limitations under the License.
  */
 
-use EngineGP\System;
-use EngineGP\Infrastructure\RemoteAccess\SshClient;
-use EngineGP\Infrastructure\RemoteAccess\SftpClient;
-use EngineGP\Infrastructure\Network\InternalIpFetcher;
-
 if (!defined('EGP')) {
     exit(header('Refresh: 0; URL=http://' . $_SERVER['HTTP_HOST'] . '/404'));
 }
@@ -39,34 +34,37 @@ class action extends actions
         $sql->query('SELECT `install` FROM `tarifs` WHERE `id`="' . $server['tarif'] . '" LIMIT 1');
         $tarif = $sql->get();
 
+        include(LIB . 'ssh.php');
+
         $sql->query('SELECT `address`, `passwd` FROM `units` WHERE `id`="' . $server['unit'] . '" LIMIT 1');
         $unit = $sql->get();
 
-        $sshClient = new SshClient($unit['address'], 'root', $unit['passwd']);
-        $sftpClient = new SftpClient($unit['address'], 'root', $unit['passwd']);
-        $internalIpFetcher = new InternalIpFetcher($sshClient);
+        // Проверка ssh соедниения пу с локацией
+        if (!$ssh->auth($unit['passwd'], $unit['address'])) {
+            return ['e' => sys::text('error', 'ssh')];
+        }
 
-        $ip = $internalIpFetcher->getInternalIp();
+        $ip = $ssh->getInternalIp();
         $port = $server['port'];
         $server_address = $server['address'] . ':' . $server['port'];
 
-        $serverSystemdStatus = trim($sshClient->execute('sudo systemctl show -p ActiveState server' . $server['uid'] . '.scope | awk -F \'=\' \'{print $2}\'', false));
+        $serverSystemdStatus = trim($ssh->get('sudo systemctl show -p ActiveState server' . $server['uid'] . '.scope | awk -F \'=\' \'{print $2}\''));
 
         if ($serverSystemdStatus == 'failed') {
-            $sshClient->execute('sudo systemctl stop server' . $server['uid'] . '.scope');
-            $sshClient->execute('sudo systemctl reset-failed server' . $server['uid'] . '.scope');
+            $ssh->set('sudo systemctl stop server' . $server['uid'] . '.scope');
+            $ssh->set('sudo systemctl reset-failed server' . $server['uid'] . '.scope');
         }
 
         // Убить процессы
-        $sshClient->execute('kill -9 `ps aux | grep s_' . $server['uid'] . ' | grep -v grep | awk ' . "'{print $2}'" . ' | xargs;'
+        $ssh->set('kill -9 `ps aux | grep s_' . $server['uid'] . ' | grep -v grep | awk ' . "'{print $2}'" . ' | xargs;'
             . 'lsof -i@' . $server_address . ' | awk ' . "'{print $2}'" . ' | grep -v PID | xargs`; sudo -u server' . $server['uid'] . ' tmux kill-session -t server' . $server['uid']);
 
         // Временный файл
-        $temp = System::temp(action::config($ip, $port, $server['slots_start'], $sshClient->execute('cat ' . $tarif['install'] . '/' . $server['uid'] . '/server.cfg', false)));
+        $temp = sys::temp(action::config($ip, $port, $server['slots_start'], $ssh->get('cat ' . $tarif['install'] . '/' . $server['uid'] . '/server.cfg')));
 
         // Обновление файла server.cfg
-        $sftpClient->putFile($temp, $tarif['install'] . $server['uid'] . '/server.cfg');
-        $sshClient->execute('chmod 0644' . ' ' . $tarif['install'] . $server['uid'] . '/server.cfg');
+        $ssh->setfile($temp, $tarif['install'] . $server['uid'] . '/server.cfg');
+        $ssh->set('chmod 0644' . ' ' . $tarif['install'] . $server['uid'] . '/server.cfg');
 
         unlink($temp);
 
@@ -74,19 +72,19 @@ class action extends actions
         $bash = './samp03svr-cr';
 
         // Временный файл
-        $temp = System::temp($bash);
+        $temp = sys::temp($bash);
 
         // Обновление файла start.sh
-        $sftpClient->putFile($temp, $tarif['install'] . $server['uid'] . '/start.sh');
-        $sshClient->execute('chmod 0500' . ' ' . $tarif['install'] . $server['uid'] . '/start.sh');
+        $ssh->setfile($temp, $tarif['install'] . $server['uid'] . '/start.sh');
+        $ssh->set('chmod 0500' . ' ' . $tarif['install'] . $server['uid'] . '/start.sh');
 
         // Строка запуска
-        $sshClient->execute('cd ' . $tarif['install'] . $server['uid'] . ';' // переход в директорию игрового сервера
+        $ssh->set('cd ' . $tarif['install'] . $server['uid'] . ';' // переход в директорию игрового сервера
             . 'rm *.pid;' // Удаление *.pid файлов
             . 'sudo -u server' . $server['uid'] . ' mkdir -p oldstart;' // Создание папки логов
             . 'cat server_log.txt >> oldstart/' . date('d.m.Y_H:i:s', $server['time_start']) . '.log; rm server_log.txt; rm oldstart/01.01.1970_03:00:00.log;'  // Перемещение лога предыдущего запуска
             . 'chown server' . $server['uid'] . ':servers server.cfg start.sh;' // Обновление владельца файлов server.cfg start.sh
-            . 'sudo systemd-run --unit=server' . $server['uid'] . ' --scope -p CPUQuota=' . $server['cpu'] . '% -p MemoryMax=' . $server['ram'] . 'M sudo -u server' . $server['uid'] . ' tmux new-session -ds s_' . $server['uid'] . ' sh -c ./start.sh'); // Запуск игрового сервера
+            . 'sudo systemd-run --unit=server' . $server['uid'] . ' --scope -p CPUQuota=' . $server['cpu'] . '% -p MemoryMax=' . $server['ram'] . 'M sudo -u server' . $server['uid'] . ' tmux new-session -ds s_' . $server['uid'] . ' sh -c ./start.sh'); // Запуск игровго сервера
 
         // Обновление информации в базе
         $sql->query('UPDATE `servers` set `status`="' . $type . '", `online`="0", `players`="", `time_start`="' . $start_point . '", `stop`="1" WHERE `id`="' . $id . '" LIMIT 1');
@@ -96,11 +94,8 @@ class action extends actions
         // Сброс кеша
         actions::clmcache($id);
 
-        System::reset_mcache('server_scan_mon_pl_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => $type, 'online' => 0, 'players' => '']);
-        System::reset_mcache('server_scan_mon_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => $type, 'online' => 0]);
-
-        $sshClient->disconnect();
-        $sftpClient->disconnect();
+        sys::reset_mcache('server_scan_mon_pl_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => $type, 'online' => 0, 'players' => '']);
+        sys::reset_mcache('server_scan_mon_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => $type, 'online' => 0]);
 
         return ['s' => 'ok'];
     }
